Added Layer threshold_relu

+++ packages/tf/layers/zero_padding_1D/zero_padding_1D.config.ts
@@ -2,14 +2,14 @@ import { Int, Str, Bool, None, Tuple } from "@/packages/typewriter";
 import { ArgsInstance } from "../base";
 import { Float } from "@/packages/typewriter/float";
 
-// tf.keras.layers.ZeroPadding1D(
-//   padding=1, **kwargs
-// // )
+// tf.keras.layers.ThresholdedReLU(
+//   theta=1.0, **kwargs
+// )
 
 export const args: ArgsInstance[] = [
 {
   name: "padding",
   isRequired: false,
-  defaultValue: Int.of(1)
+  defaultValue: Float.of(1.0),
 }
 ];
